+++ CHANGES.rst
@@ -43,6 +43,9 @@ Changelog
 
 - Rewrite ``ligo-skymap-plot-stats`` to reduce code duplication.
 
+- Add ``--measurement-error gaussian-noise`` option to
+  ``bayestar-realize-coincs`` to simulate a matched filter in Gaussian noise.
+
 0.0.12 (2018-07-18)
 ===================